refactor: managing actions in a specific component
